Adopt the Amazon Open Source Code of Conduct. (#1)
Content copy-pasted from https://aws.github.io/code-of-conduct and manually formatted.  The only changes are to eliminate references to Amazon.

I picked this CoC after reviewing quite a few variants of the Contributor Covenant.  On the whole I liked the changes Amazon made, so I took this as a reasonable starting point.

Part of https://github.com/ion-fusion/fusion-java/issues/5